refactor: break storage to app-actions back-edge and dedupe export QR prep
Move maybeSave/maybeSaveAs into app-actions so the storage manager no
longer imports app-layer code. The save-first continuation moves to
DialogOrchestrator via boolean returns from handleSaveToServer and
handleSaveAsArchive, removing the pendingSaveFirst dialog state. Extract
prepareExportQrCode and reuse it in handleExport and the rack context
export, fixing the missing null-share-url guard. Drop the unused
session.ts duplicate, the unused SessionLoadResult barrel re-export, and
the handleSaveFailure export. Update stale comments in App.svelte and
dialogs.svelte.ts.

Refs #2027
